+++ Sources/KeyboardKit/Extensions/EdgeInsets+Insets.swift
@@ -11,16 +11,7 @@ import SwiftUI
 
 extension EdgeInsets {
     
-    /**
-     Create an `EdgeInsets` with the same insets everywhere.
-     */
-    static func all(_ all: CGFloat) -> EdgeInsets {
-        self.init(top: all, leading: all, bottom: all, trailing: all)
-    }
-    
-    /**
-     Create an `EdgeInsets` with horizontal/vertical values.
-     */
+    /// Create a value with horizontal/vertical values.
     static func horizontal(_ horizontal: CGFloat, vertical: CGFloat) -> EdgeInsets {
         self.init(top: vertical, leading: horizontal, bottom: vertical, trailing: horizontal)
     }